Improve handling of LLVM dylib build
If Clang is linked against LLVM dynamically, then we risk
linking LLVM twice if we use static linking ourselves.

Previously, there was a workaround removing the shared lib
dependency from the imported LLVM target.
This workaround however depends on a certain LLVM build configuration,
as it only overrides the target for the RelWithDebInfo build type.