Clang extractor: don't consume tokens early for decls
When a decl is referenced, we build the info for the decl.
This can lead to incorrect token annotations, for example for recursive functions.
When the recursive call happens, the info for the whole function is constructed (because the func is referenced).
Previously, this led to all remaining tokens being attributed to the whole function decl.